backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.1.8
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
Flask==3.1.3
flask-cors==6.0.2
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
multidict==6.7.1
openai==2.24.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-telegram-bot==20.7
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
yarl==1.22.0
zipp==3.23.0
